rename components and variables
